Move symlink test helper to test package
